services/api-gateway/requirements.txt
fastapi==0.115.12
uvicorn==0.32.1
pydantic==2.10.0
firebase-admin==6.6.0
google-cloud-firestore==2.19.0
google-cloud-tasks==2.16.0
httpx==0.28.1
google-auth==2.38.0
